+++ components/camel-mongodb-gridfs/src/test/java/org/apache/camel/component/mongodb/gridfs/integration/GridFsProducerOperationsIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.mongodb.gridfs;
+package org.apache.camel.component.mongodb.gridfs.integration;
 
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
@@ -24,6 +24,8 @@ import java.util.Map;
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.mongodb.gridfs.GridFsConstants;
+import org.apache.camel.component.mongodb.gridfs.GridFsEndpoint;
 import org.bson.types.ObjectId;
 import org.junit.jupiter.api.Test;
 
@@ -33,7 +35,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class GridFsProducerOperationsTest extends AbstractMongoDbTest {
+public class GridFsProducerOperationsIT extends AbstractMongoDbITSupport {
 
     @Override
     protected RouteBuilder createRouteBuilder() throws Exception {
